Only "pay for shopping cart" left Tests not passed : 4 3 for pay for cart 1 for update product

+++ src/main/java/sk/stuba/fei/uim/oop/assignment3/productlogic/ProductAmountResponse.java
@@ -8,8 +8,8 @@ import lombok.Setter;
 @Getter
 @Setter
 public class ProductAmountResponse {
-    int amount;
-    public ProductAmountResponse(int amount){
+    long amount;
+    public ProductAmountResponse(long amount){
         this.amount = amount;
     }
